app/requirements.txt
urllib3==1.26.4
requests==2.25.1
PyYAML==5.4.1
